Fixed a bug.
When a user was not logged in and clicked a book in the book search table they were taken to a registration page. The issue was that for some reason the 'already have an account' button was sending them to books/login instead of account/login. Couldn't find a logical reason but fixed the issue anyway.