Fixed a bug where users could borrow their own books.
The book details page now has two modals which can show. If the logged in users is the same as the user who has uploaded the book who's details are currently being viewed then an error modal will show.
Otherwise the normal contact modal will show and borrowing will go on as normal.